Change of number navigation bar
I have decided later on to keep the navigation as three as it would be too much on the navigation bar and much less information.

I had planned on creating 4 as  a vision page as page 3 was supposed to be for a mission. Page 5 would have been the synopsis but give that there are some limitations

- lack of time
- learning progress might slow down a bit
-essential  time wasted

I chose to keep it to three as I hope it would be more effective, cleaner, and better put out